Split face.c into face.c and dynamicface.c
- Created dynamicface.h for dynamic face operations
- Created dynamicface.c containing functions prefixed with dynamic
- Updated face.c to contain only initialization and utility functions
- Renamed countEdge to dynamicCountEdge for consistency
- Updated Makefile to include the new files
- Moved CycleForcedCounter and CycleSetReducedCounter to dynamicface.c
